Create decks/add.rs
